testproject/requirements.txt
face_recognition_models
Click>=6.0
numpy
Pillow
cmake
https://files.pythonhosted.org/packages/0e/ce/f8a3cff33ac03a8219768f0694c5d703c8e037e6aba2e865f9bae22ed63c/dlib-19.8.1-cp36-cp36m-win_amd64.whl#sha256=794994fa2c54e7776659fddb148363a5556468a6d5d46be8dad311722d54bfcf
opencv-python
face_recognition
matplotlib
h5py
keras==1.2.2
tensorflow==1.7.0
scipy==1.0.0
django
djangorestframework
django-filter